requirements.txt
asgiref==3.6.0
Brlapi==0.8.2
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
cupshelpers==1.0
dbus-python==1.2.16
distlib==0.3.6
distro==1.5.0
distro-info==1.0
Django==4.1.7
django-parler==2.3
django-parler-rest==2.2
django-rosetta==0.9.9
djangorestframework==3.14.0
filelock==3.9.0
httplib2==0.18.1
idna==2.10
louis==3.16.0
Mako==1.1.3
MarkupSafe==1.1.1
platformdirs==3.1.1
polib==1.2.0
pycairo==1.16.2
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
pytz==2022.7.1
pyxdg==0.27
reportbug==7.10.3+deb11u1
requests==2.25.1
six==1.16.0
sqlparse==0.4.3
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.20.0
xdg==5
